fix: move tests to correct file
